requirements.txt
Django==5.0.7
djangorestframework==3.15.2
django-filter==24.2
djangorestframework-simplejwt==5.3.1
dj-rest-auth==6.0.0
django-allauth==0.59.0
psycopg[binary]==3.2.10
python-dotenv==1.0.1
dj-database-url==2.2.0
whitenoise==6.6.0
django-crispy-forms==2.1
django-axes==6.0.0
Pillow==10.4.0
pytest==8.3.2
pytest-django==4.8.0
